% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalization.R
\name{background_correction}
\alias{background_correction}
\title{Correct OD or fluorescence with data from blank wells}
\usage{
background_correction(df, .var, column_suffix)
}
\arguments{
\item{df}{A plate reader data.frame.}

\item{.var}{The unquoted variable to be background-corrected.}

\item{column_suffix}{The suffix that will be appended to the column name of the background-corrected variable.}
}
\value{
The same data.frame as the input but with an extra column with the background corrected variable
}
\description{
Function to extract the background measurements from each plate and subtract it from the samples' measurements.

function  to extract the background measurements from each plate and subtract it from the actual measurements
}
\examples{
background_correction(df = df.od, .var = OD_730, column_suffix = "bc")

}
