% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyanoInterlabShinyApp-package.R
\docType{package}
\name{CyanoInterlabShinyApp-package}
\alias{CyanoInterlabShinyApp}
\alias{CyanoInterlabShinyApp-package}
\title{CyanoInterlabShinyApp: Cyano Interlab study Shiny App}
\description{
A Shiny app to analyze the data produced in the Cyano Interlab study.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hugo-pH/CyanoInterlabShinyApp.git}
  \item Report bugs at \url{https://github.com/hugo-pH/CyanoInterlabShinyApp.git/issues}
}

}
\author{
\strong{Maintainer}: Hugo Hugo Pineda Hernandez \email{hugo.pineda.h@gmail.com}

}
\keyword{internal}
