% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalization.R
\name{od_correction}
\alias{od_correction}
\title{Normalize plate reader fluorescence values with OD}
\usage{
od_correction(
  df.od,
  df.fl,
  od_col = OD_730,
  fl_col = fl,
  bc_suffix = "bc",
  background_correction = T
)
}
\arguments{
\item{df.od}{The plate reader OD data.frame.}

\item{df.fl}{The plate reader fluorescence data.frame.}

\item{od_col}{The unqouted column name containing the OD values.}

\item{fl_col}{The unqouted column name containing the fluorescence values.}

\item{bc_suffix}{The suffix that will be appended to the column name of the background-corrected variable when \code{background_correction} = T}

\item{background_correction}{Logical indicating if background correction should be performed.}
}
\value{
A data.frame containing a new column with fluorescence per OD values.
}
\description{
Function to normalize fluorescence values by corresponding OD values in plate reader data.
}
