% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalization.R
\name{plate_reader_normalization_pipeline}
\alias{plate_reader_normalization_pipeline}
\title{Run plate reader normalization pipeline}
\usage{
plate_reader_normalization_pipeline(
  df.od,
  df.fl,
  bc_suffix,
  od_col = OD_730,
  fl_col = fl,
  background_correction = T,
  .norm_var,
  ref_strain,
  ref_induction,
  norm_suffix
)
}
\arguments{
\item{df.od}{The plate reader OD data.frame.}

\item{df.fl}{The plate reader fluorescence data.frame.}

\item{bc_suffix}{The suffix that will be appended to the column name of the background-corrected variable.}

\item{od_col}{The unqouted column name containing the OD values.}

\item{fl_col}{The unqouted column name containing the fluorescence values.}

\item{background_correction}{Logical indicating if background correction should be performed.}

\item{.norm_var}{The unquoted column name of the variable to be normalized (i.e. fl_od).}

\item{ref_strain}{The reference strain to normalize with.}

\item{ref_induction}{The reference induction condition to normalize with.}

\item{norm_suffix}{The suffix that will be appended to the column name of the normalized variable.}
}
\value{
A single data.frame containing new columns for background-corrected, fluorescence per OD and reference strain normalized data.
}
\description{
Function to combine all the background correction and normalization steps
}
\examples{
df.plate.reader <- plate_reader_normalization_pipeline(
df.od = load_plate_reader_od(df.experiments.raw),
df.fl = load_plate_reader_fl(df.experiments.raw),
bc_suffix = "bc",
.norm_var = fl_od,
ref_strain = "J23100",
ref_induction = "-",
norm_suffix  = "norm")

  
}
