% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_summary}
\alias{plot_summary}
\title{Summary plot}
\usage{
plot_summary(
  df,
  y_var,
  colour_var,
  f_rows_var,
  f_cols_var,
  df.labels,
  log_space,
  free_y_scale = F
)
}
\arguments{
\item{df}{A data.frame containing the data to be plotted.}

\item{y_var}{A string with the column name of the variable for the y axis.}

\item{colour_var}{A string with the column name of the variable to determine the colours.}

\item{f_rows_var}{A string with the column name of the variable to the determine the rows for the facet_grid.}

\item{f_cols_var}{A string with the column name of the variable to the determine the columns for the facet_grid.}

\item{df.labels}{A data.frame with plot labels.}
}
\value{

}
\description{
A function to produce a summary plot of a given y variable, chossing with varibles will be used for colours and for facets
}
